requirements.txt
azure
azureml-sdk==1.0.76
azureml-widgets
joblib==0.14.0
jupyter
python-dotenv
numpy
scikit-learn==0.22
matplotlib==3.1.2
scikit-plot
keras==2.2.5
tensorflow==1.13.1